Refactor for Unit Testing Clean Project
