--- app/src/main/java/com/lonedev/moviecatalogue/base/observable/BaseObservable.kt
@@ -1,8 +1,0 @@
-/*
- * Created by Bagus Yogatama on 8/28/19 9:52 PM
- * Copyright (c) 2019 . All rights reserved.
- * Last modified 8/28/19 9:52 PM
- */
-
-package com.lonedev.moviecatalogue.base.observable
-
