+++ app/src/main/java/com/lonedev/moviecatalogue/di/modules/AppModule.kt
@@ -14,7 +14,7 @@ import dagger.Provides
 import javax.inject.Singleton
 
 @Module(includes = [NetworkModule::class, ViewModelModule::class, DatabaseModule::class])
-class AppModule {
+object AppModule {
 
     @Singleton
     @Provides
