+++ app/src/main/res/layout/activity_detail_movies.xml
@@ -166,7 +166,7 @@
                 <TextView
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
-                        android:text="Videos"
+                        android:text="@string/videos"
                         android:textStyle="bold"
                         android:textSize="18sp"
                         android:textColor="#000000"/>
