+++ app/src/main/res/values/strings.xml
@@ -26,4 +26,5 @@
     <string name="no_tv_released">No TV Series Released Today</string>
     <string name="no_movie_released">No Movies Released Today</string>
     <string name="general_error">Something went wrong, please try again later</string>
+    <string name="videos">Videos</string>
 </resources>
