+++ app/src/test/java/com/lonedev/moviecatalogue/TVSeriesFragmentTest.kt
@@ -46,11 +46,11 @@ class TVSeriesFragmentTest {
             tvSeriesReporsitory,
             TestSchedulerProvider(testScheduler)
         )
-        tvSeriesViewModel.getTVSeries()
     }
 
     @Test
     fun getTVSeries() {
+        tvSeriesViewModel.getTVSeries()
         tvSeriesViewModel.onSuccessGetTVSeries().observeForever {
             it.isNotEmpty()
         }
